# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "马森雨个人技术博客（masenyu.top）RESTful API 接口文档，涵盖文章、分类、标签、项目、友链、评论、AI助手、数据看板与站点管理等模块。",
        "title": "马森雨个人技术博客 API",
        "contact": {
            "name": "马森雨",
            "url": "https://masenyu.top"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/admin/ai/conversations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "查询当前管理员创建的全部 AI 对话会话",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI 助手与智能工具"
                ],
                "summary": "获取 AI 对话会话列表",
                "responses": {
                    "200": {
                        "description": "会话列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handler.AIConversationDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "创建一个空对话会话，随后可调用流式对话接口交流",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI 助手与智能工具"
                ],
                "summary": "新建 AI 对话会话",
                "responses": {
                    "200": {
                        "description": "新建会话",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.AIConversationDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "批量删除当前管理员名下的全部对话记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI 助手与智能工具"
                ],
                "summary": "清空全部 AI 对话会话",
                "responses": {
                    "200": {
                        "description": "清空成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/ai/conversations/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "根据会话 ID 获取会话元数据及完整的历史消息记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI 助手与智能工具"
                ],
                "summary": "获取指定 AI 会话详情与历史消息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "会话详情",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.AIConversationDetailDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "会话不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "根据会话 ID 删除对话及其所有消息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI 助手与智能工具"
                ],
                "summary": "删除单个 AI 对话会话",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "会话不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "修改会话标题名称",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI 助手与智能工具"
                ],
                "summary": "重命名 AI 对话会话",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "新标题",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ConversationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.AIConversationDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "会话不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/ai/conversations/{id}/messages/stream": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "向指定会话发送提问，并通过 Server-Sent Events (SSE) 持续接收流式输出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "AI 助手与智能工具"
                ],
                "summary": "发送消息并流式接收 AI 回复 (SSE)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "消息内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ConversationMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE 流式事件 (meta/delta/done/error)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "会话不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AIConversationDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastMessageAt": {
                    "type": "string"
                },
                "messageCount": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "titleMode": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "handler.AIConversationDetailDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastMessageAt": {
                    "type": "string"
                },
                "messageCount": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.AIMessageDTO"
                    }
                },
                "model": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "titleMode": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "handler.AIMessageDTO": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "errorMessage": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "sequence": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handler.ConversationMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "handler.ConversationRequest": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "参数错误"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "请输入 JWT Token，格式为: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CsrfToken": {
            "description": "管理端防跨站请求伪造令牌",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}